requirements.txt
Django>=4.0.0
django-ratelimit>=4.1.0
Pillow>=9.0.0
django-encrypted-model-fields>=0.6.0
python-dotenv>=0.19.0 